skipped plan to use async IO as there is no competing IO
Main IO is reading from and writing to UDP socket. Real driving controllers may want to write to disk for logging, but in general, this must be done synchronized to the network IO, not in parallel. If there is real number crunching to happen in parallel, multithreading is an option since the GIL will be released by socket operations.